made the redis cache completely optional and async
If the REDIS_URL is missing, the cache is not created.

All redis interactions are now async.
The downside of this is that it can't be easily used with diesel
since diesel would block the actix runtime.

So caching is now only used for the duckduckgo API.
Once diesel supports async, we can start caching again.

Signed-off-by: Bart Willems <bwillems@protonmail.com>